Rename relative_position_sensor
