+++ s2e-ff/src/components/aocs/initialize_relative_position_sensor.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "RelativePositionSensor.hpp"
+#include "relative_position_sensor.hpp"
 
 RelativePositionSensor InitializeRelativePositionSensor(ClockGenerator* clock_gen, const std::string file_name, const double compo_step_time_s,
                                                         const RelativeInformation& rel_info, const Dynamics& dynamics,
