+++ s2e-ff/src/components/aocs/relative_position_sensor.cpp
@@ -1,4 +1,4 @@
-#include "RelativePositionSensor.hpp"
+#include "relative_position_sensor.hpp"
 
 RelativePositionSensor::RelativePositionSensor(const int prescaler, ClockGenerator* clock_gen, Sensor& sensor_base, const int target_sat_id,
                                                const int reference_sat_id, const RelativePositionSensorErrorFrame error_frame,
