+++ s2e-ff/src/simulation/spacecraft/ff_components_2.hpp
@@ -8,8 +8,8 @@
 #include <components/ideal/force_generator.hpp>
 #include <components/real/cdh/on_board_computer.hpp>
 
-#include "../../components/aocs/InitializeRelativePositionSensor.hpp"
 #include "../../components/aocs/initialize_relative_distance_sensor.hpp"
+#include "../../components/aocs/initialize_relative_position_sensor.hpp"
 #include "../../components/ideal/InitializeRelativeAttitudeController.hpp"
 
 class FfComponents2 : public InstalledComponents {
